refactor: improve email module templates
Email Module templates now have a dedicated place inside application layer inside email module.

+++ src/users/application/services/user.service.ts
@@ -75,6 +75,10 @@ export class UserService {
     }
 
     public async deleteUser(id: string): Promise<string> {
+        const user = await this.findOneUser(id);
+
+        if (!user) return null;
+        
         const res: UserEntity = await this.repository.delete(id);
 
         return res ? 'User deleted' : undefined;
